+++ lib/hotwire_native_version_gate/concern.rb
@@ -9,12 +9,12 @@ module HotwireNativeVersionGate
         VersionGate.native_feature(feature, ios: ios, android: android)
       end
 
-      def native_version_regexes=(regexes)
-        VersionGate.native_version_regexes = regexes
+      def native_version_regexes
+        VersionGate.native_version_regexes
       end
 
-      def prepend_native_version_regexes(regexes)
-        VersionGate.prepend_native_version_regexes(regexes)
+      def native_version_regexes=(regexes)
+        VersionGate.native_version_regexes = regexes
       end
     end